+++ haproxy-minio/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Version bump for new base image
 * Add tune.disable-zero-copy-forwarding to haproxy global section to deal with CPU overusage
+* Allow for 1-4 minio servers
 
 ---